requirements.txt
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.7
django-admin-sortable2==0.7.6
django-unused-media==0.2.0
html5lib==1.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
Pyphen==0.9.5
pytz==2019.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
